+++ example/lib/books/src/screens/authors.dart
@@ -3,9 +3,9 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
 
 import '../data/library.dart';
-import '../routing.dart';
 import '../widgets/author_list.dart';
 
 class AuthorsScreen extends StatelessWidget {