+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 0.8.3
+- Changed `debugOutputFullPaths` to `debugLogDiagnostics` and added add'l
+  debugging logging
+- TODO: fix `context.go` and `Link.followLink` in `settings.dart`
+- TODO: pass state to redirect to package up `location` and `params` to enable
+  redirection of '/books/:bookId' => '/books/all/${state.params['bookId']}'
+
+
 ### 0.8.2
 - updated README for `Link` widget support